LayoutTests/fast/html/details-summary-document-child.html

<html>
<body onload="test()">
<script>
if (window.testRunner)
    testRunner.dumpAsText();

function test() {
    document.open();
    document.insertBefore(document.createElement("summary"), null);
    // The document don't have <body> so we need use console.log()
    // to make readable expectation.
    console.log("PASS unless crash");
}
</script>
</body>
</html>
