LayoutTests/fast/html/details-summary-document-child-expected.txt

CONSOLE MESSAGE: line 12: PASS unless crash
